# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/theaters": {
            "get": {
                "tags": [
                    "Theaters"
                ],
                "summary": "Lấy danh sách rạp chiếu",
                "operationId": "getTheatersList",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "total": 50,
                                                "per_page": 15,
                                                "current_page": 1,
                                                "last_page": 4,
                                                "first_page_url": "http://cyphercinema.com?page=1",
                                                "last_page_url": "http://cyphercinema.com?page=4",
                                                "next_page_url": "http://cyphercinema.com?page=2",
                                                "prev_page_url": null,
                                                "path": "http://cyphercinema.com",
                                                "from": 1,
                                                "to": 15,
                                                "data": [
                                                    {
                                                        "id": 1,
                                                        "name": "BHD STAR DISCOVERY",
                                                        "image": "https://image.com/1",
                                                        "address": "Tầng 8, TTTM Discovery – 302 Cầu Giấy, P.Dịch Vọng, Quận Cầu Giấy, Hà Nội",
                                                        "content": "CÁC QUY ĐỊNH GIÁ VÉ \n – Giá vé trẻ em áp dụng cho trẻ em có chiều cao dưới 1,3m. Yêu cầu trẻ em có mặt khi mua vé. Trẻ em dưới 0,7m sẽ được miễn phí vé khi mua cùng 01 vé người lớn đi kèm theo. Không áp dụng kèm với chương trình khuyến mãi ưu đãi về giá vé khác. \n – Giá vé thành viên U22 chỉ áp dụng cho thành viên dưới 22 tuổi khi mua vé. Không áp dụng kèm với chương trình khuyến mãi ưu đãi về giá vé khác. Mỗi thẻ thành viên U22 được áp dụng giá vé ưu đãi tối đa 02 vé/ngày. \n – Ngày lễ: 1/1, Giổ Tổ Hùng Vương 10/3 Âm Lịch, 30/4, 1/5, 02 ngày Lễ Quốc Khánh \n – Giá vé Tết Âm Lịch sẽ được áp dụng riêng. \n – Suất chiếu đặc biệt áp dụng giá vé theo khung giờ của ngày. Không áp dụng các giá vé ưu đãi dành cho U22, Privilege Voucher/Staff Voucher, Happy Day. Trong trường hợp Suất chiếu đặc biệt cùng ngày với Happy Day sẽ áp dụng giá vé của Thứ 3",
                                                        "city": "Hà Nội",
                                                        "phone": "098765432111",
                                                        "email": "bhd@gmail.com"
                                                    },
                                                    {
                                                        "id": 2,
                                                        "name": "BHD STAR PHAM HUNG",
                                                        "image": "https://image.com/2",
                                                        "address": "Tầng 8, TTTM Discovery – 302 Cầu Giấy, P.Dịch Vọng, Quận Cầu Giấy, Hà Nội",
                                                        "content": "CÁC QUY ĐỊNH GIÁ VÉ \n – Giá vé trẻ em áp dụng cho trẻ em có chiều cao dưới 1,3m. Yêu cầu trẻ em có mặt khi mua vé. Trẻ em dưới 0,7m sẽ được miễn phí vé khi mua cùng 01 vé người lớn đi kèm theo. Không áp dụng kèm với chương trình khuyến mãi ưu đãi về giá vé khác. \n – Giá vé thành viên U22 chỉ áp dụng cho thành viên dưới 22 tuổi khi mua vé. Không áp dụng kèm với chương trình khuyến mãi ưu đãi về giá vé khác. Mỗi thẻ thành viên U22 được áp dụng giá vé ưu đãi tối đa 02 vé/ngày. \n – Ngày lễ: 1/1, Giổ Tổ Hùng Vương 10/3 Âm Lịch, 30/4, 1/5, 02 ngày Lễ Quốc Khánh \n – Giá vé Tết Âm Lịch sẽ được áp dụng riêng. \n – Suất chiếu đặc biệt áp dụng giá vé theo khung giờ của ngày. Không áp dụng các giá vé ưu đãi dành cho U22, Privilege Voucher/Staff Voucher, Happy Day. Trong trường hợp Suất chiếu đặc biệt cùng ngày với Happy Day sẽ áp dụng giá vé của Thứ 3",
                                                        "city": "Hà Nội",
                                                        "phone": "098765432111",
                                                        "email": "bhd@gmail.com"
                                                    }
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Theaters"
                ],
                "summary": "Tạo rạp chiếu mới",
                "operationId": "createTheater",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Rạp chiếu phim Homeless"
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "example": 29
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "TƯNG BỪNG KHAI TRƯƠNG RẠP CHIẾU PHIM BHD STAR THE GARDEN: RỘN RÀNG CÙNG CHUỖI ƯU ĐÃI SIÊU KHỦNG. Cụm rạp thứ 10 của BHD Star Cineplex sẽ chính thức khai trương vào ngày 21/12/2019 tại Hà Nội với ưu đãi xem phim miễn phí cùng hàng ngàn chương trình khuyến mãi siêu khủng!"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "Poster cho bài viết sự kiện"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "bha@mama.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0987654321234"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Tầng 8, TTTM Discovery – 302 Cầu Giấy, P.Dịch Vọng, Quận Cầu Giấy, Hà Nội"
                                    }
                                },
                                "required": ["name", "city_id", "content", "image", "email", "phone", "address"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Voucher created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "BHD STAR DISCOVERY",
                                                "image": "https://image.com/1",
                                                "address": "Tầng 8, TTTM Discovery – 302 Cầu Giấy, P.Dịch Vọng, Quận Cầu Giấy, Hà Nội",
                                                "content": "CÁC QUY ĐỊNH GIÁ VÉ \n – Giá vé trẻ em áp dụng cho trẻ em có chiều cao dưới 1,3m. Yêu cầu trẻ em có mặt khi mua vé. Trẻ em dưới 0,7m sẽ được miễn phí vé khi mua cùng 01 vé người lớn đi kèm theo. Không áp dụng kèm với chương trình khuyến mãi ưu đãi về giá vé khác. \n – Giá vé thành viên U22 chỉ áp dụng cho thành viên dưới 22 tuổi khi mua vé. Không áp dụng kèm với chương trình khuyến mãi ưu đãi về giá vé khác. Mỗi thẻ thành viên U22 được áp dụng giá vé ưu đãi tối đa 02 vé/ngày. \n – Ngày lễ: 1/1, Giổ Tổ Hùng Vương 10/3 Âm Lịch, 30/4, 1/5, 02 ngày Lễ Quốc Khánh \n – Giá vé Tết Âm Lịch sẽ được áp dụng riêng. \n – Suất chiếu đặc biệt áp dụng giá vé theo khung giờ của ngày. Không áp dụng các giá vé ưu đãi dành cho U22, Privilege Voucher/Staff Voucher, Happy Day. Trong trường hợp Suất chiếu đặc biệt cùng ngày với Happy Day sẽ áp dụng giá vé của Thứ 3",
                                                "city": "Hà Nội",
                                                "phone": "098765432111",
                                                "email": "bhd@gmail.com"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/theaters/{id}": {
            "get": {
                "tags": [
                    "Theaters"
                ],
                "summary": "Lấy rạp chiếu theo ID",
                "operationId": "getTheaterById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A single voucher",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "BHD STAR DISCOVERY",
                                                "image": "https://image.com/1",
                                                "address": "Tầng 8, TTTM Discovery – 302 Cầu Giấy, P.Dịch Vọng, Quận Cầu Giấy, Hà Nội",
                                                "content": "CÁC QUY ĐỊNH GIÁ VÉ \n – Giá vé trẻ em áp dụng cho trẻ em có chiều cao dưới 1,3m. Yêu cầu trẻ em có mặt khi mua vé. Trẻ em dưới 0,7m sẽ được miễn phí vé khi mua cùng 01 vé người lớn đi kèm theo. Không áp dụng kèm với chương trình khuyến mãi ưu đãi về giá vé khác. \n – Giá vé thành viên U22 chỉ áp dụng cho thành viên dưới 22 tuổi khi mua vé. Không áp dụng kèm với chương trình khuyến mãi ưu đãi về giá vé khác. Mỗi thẻ thành viên U22 được áp dụng giá vé ưu đãi tối đa 02 vé/ngày. \n – Ngày lễ: 1/1, Giổ Tổ Hùng Vương 10/3 Âm Lịch, 30/4, 1/5, 02 ngày Lễ Quốc Khánh \n – Giá vé Tết Âm Lịch sẽ được áp dụng riêng. \n – Suất chiếu đặc biệt áp dụng giá vé theo khung giờ của ngày. Không áp dụng các giá vé ưu đãi dành cho U22, Privilege Voucher/Staff Voucher, Happy Day. Trong trường hợp Suất chiếu đặc biệt cùng ngày với Happy Day sẽ áp dụng giá vé của Thứ 3",
                                                "city": "Hà Nội",
                                                "phone": "098765432111",
                                                "email": "bhd@gmail.com"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Theaters"
                ],
                "summary": "Chỉnh sửa rạp chiếu",
                "operationId": "editTheater",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Rạp chiếu phim Homeless"
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "example": 29
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "TƯNG BỪNG KHAI TRƯƠNG RẠP CHIẾU PHIM BHD STAR THE GARDEN: RỘN RÀNG CÙNG CHUỖI ƯU ĐÃI SIÊU KHỦNG. Cụm rạp thứ 10 của BHD Star Cineplex sẽ chính thức khai trương vào ngày 21/12/2019 tại Hà Nội với ưu đãi xem phim miễn phí cùng hàng ngàn chương trình khuyến mãi siêu khủng!"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "Poster cho bài viết sự kiện"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "bha@mama.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0987654321234"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Tầng 8, TTTM Discovery – 302 Cầu Giấy, P.Dịch Vọng, Quận Cầu Giấy, Hà Nội"
                                    }
                                },
                                "required": ["name", "city_id", "content", "image", "email", "phone", "address"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật rạp chiếu thành công.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "BHD STAR DISCOVERY",
                                                "image": "https://image.com/1",
                                                "address": "Tầng 8, TTTM Discovery – 302 Cầu Giấy, P.Dịch Vọng, Quận Cầu Giấy, Hà Nội",
                                                "content": "CÁC QUY ĐỊNH GIÁ VÉ \n – Giá vé trẻ em áp dụng cho trẻ em có chiều cao dưới 1,3m. Yêu cầu trẻ em có mặt khi mua vé. Trẻ em dưới 0,7m sẽ được miễn phí vé khi mua cùng 01 vé người lớn đi kèm theo. Không áp dụng kèm với chương trình khuyến mãi ưu đãi về giá vé khác. \n – Giá vé thành viên U22 chỉ áp dụng cho thành viên dưới 22 tuổi khi mua vé. Không áp dụng kèm với chương trình khuyến mãi ưu đãi về giá vé khác. Mỗi thẻ thành viên U22 được áp dụng giá vé ưu đãi tối đa 02 vé/ngày. \n – Ngày lễ: 1/1, Giổ Tổ Hùng Vương 10/3 Âm Lịch, 30/4, 1/5, 02 ngày Lễ Quốc Khánh \n – Giá vé Tết Âm Lịch sẽ được áp dụng riêng. \n – Suất chiếu đặc biệt áp dụng giá vé theo khung giờ của ngày. Không áp dụng các giá vé ưu đãi dành cho U22, Privilege Voucher/Staff Voucher, Happy Day. Trong trường hợp Suất chiếu đặc biệt cùng ngày với Happy Day sẽ áp dụng giá vé của Thứ 3",
                                                "city": "Hà Nội",
                                                "phone": "098765432111",
                                                "email": "bhd@gmail.com"
                                            }
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Cập nhật rạp chiếu thành công."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy sự kiện",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Không tìm thấy ID của rạp."
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Theaters"
                ],
                "summary": "Xóa rạp chiếu",
                "operationId": "deleteTheater",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "ID của rạp chiếu cần xóa"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rạp đã được xóa thành công.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": null,
                                            "example": null
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Rạp đã được xóa thành công."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy rạp",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Không tìm thấy ID của rạp."
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}